+++ SLK_img2pixel/src/process.h
@@ -32,6 +32,7 @@ extern int contrast;
 extern int img_gamma;
 extern int saturation;
 extern int dither_amount;
+extern int alpha_threshold;
 
 void process_image(const SLK_RGB_sprite *in, SLK_RGB_sprite *out, SLK_Palette *palette, int sample_mode, int process_mode);