+++ SLK_img2pixel/changelog.md
@@ -9,6 +9,9 @@ Version:                SLK_img2pixel 1.0 - initial release (tba)
 
 CHANGES:
 
+* added bilinear sampling mode
+* added alpha threshold
+
 ----------------------------------------
 Version:                SLK_img2pixel 0.1 - initial prerelease (20.01.2021)
 ----------------------------------------